pyproject.toml
[tool.poetry]
name = "printprob-sheets-automation"
version = "0.1.0"
description = "Automation of writes to Google sheets"
authors = ["Sriram Viswanathan <sriramvi@andrew.cmu.edu>"]
packages = [
    { include = "sheets" },
    { include = "sheets/**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
pygsheets = "^2.0.5"
numpy = "^1.23.2"
google-api-python-client = "^2.58.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chewfiles = "sheets.cli:main"